+++ Vectoring/GeneralObjectData.cs
@@ -7,8 +7,8 @@ using Logging;
 public class RoomObj<T> where T : class
 {
 	public T Value;
-	public byte X, Y;
-	public RoomObj(T worldObj, byte X, byte Y)
+	public int X, Y;
+	public RoomObj(T worldObj, int X, int Y)
 	{
 		Value = worldObj;
 		this.X = X;
@@ -23,7 +23,7 @@ public class RoomObj<T> where T : class
 public class WorldObj<T> : RoomObj<T> where T : class
 {
 	public byte Room;
-	public WorldObj(T worldObj, byte X, byte Y, byte Room) : base(worldObj, X, Y)
+	public WorldObj(T worldObj, int X, int Y, byte Room) : base(worldObj, X, Y)
 	{
 		this.Room = Room;
 	}
